add: Handle text message and initial structure to filter commands
The filter function when a text message arrive use a tuple to decide if
the message is a command registered. Validated with regex (just the
command at this point) select the function will be executed according to
command. And by default respond with generic function. Later will be
implemented commands with parameters